adding vendors queries and mutations

+++ src/graphql/products/products.resolver.ts
@@ -1,6 +1,6 @@
 import { Resolver, Query, Mutation, Args, Context, Subscription } from '@nestjs/graphql';
-import { UnauthorizedException } from '@nestjs/common';
-import { PubSub } from 'graphql-subscriptions';
+import { UnauthorizedException, NotFoundException } from '@nestjs/common';
+// import { PubSub } from 'graphql-subscriptions';
 
 import { ProductType } from './dto/product.dto';
 import { ProductInput, ProductFilter, SearchQueryInput } from './dto/product.input';
@@ -8,7 +8,7 @@ import { ProductsService } from '../../shared/services/products.service';
 import { IUser } from '../../shared/interfaces/user.interface';
 import { WatsonDicoveryService } from '../../shared/services/watson-discovery.service';
 
-const pubSub = new PubSub();
+// const pubSub = new PubSub();
 
 @Resolver('Products')
 export class ProductsResolver {
@@ -25,7 +25,7 @@ export class ProductsResolver {
     const response = await this.watsonDiscoveryService.queryCollection(searchQueryInput);
     const ids = response.result.results.map(x => x.id);
     
-    return this.productsService.products({
+    return await this.productsService.allProducts({
       ibmDiscoveryDocumentId: { $in: ids },
       status: 'approved',
       archived: false
@@ -33,15 +33,15 @@ export class ProductsResolver {
   }
 
   @Query(() => ProductType)
-  async getProductByName(
+  async productByName(
     @Context('user') user: IUser,
     @Args('productName') productName: string,
   ): Promise<ProductType> {
-    return this.productsService.getProductByName(productName);
+    return await this.productsService.productByName(productName);
   }
 
   @Query(() => [ProductType])
-  async products(
+  async allProducts(
     @Context('user') user: IUser,
     @Args('filter') filter: ProductFilter,
   ): Promise<ProductType[]> {
@@ -49,11 +49,11 @@ export class ProductsResolver {
       throw new UnauthorizedException('You do not have the permission to retrieve products');
     }
     
-    return this.productsService.products(filter);
+    return await this.productsService.allProducts(filter);
   }
 
   @Query(() => ProductType)
-  async getProductById(
+  async productById(
     @Context('user') user: IUser,
     @Args('productId') productId: string
   ) {
@@ -61,7 +61,7 @@ export class ProductsResolver {
       throw new UnauthorizedException('You do not have the permission to retrieve products');
     }
 
-    return this.productsService.getProductById(productId);
+    return await this.productsService.productById(productId);
   }
 
   @Mutation(() => ProductType)
@@ -73,7 +73,7 @@ export class ProductsResolver {
       throw new UnauthorizedException('You do not have the permission to create a product');
     }
 
-    return this.productsService.addProduct(user.sub, product);
+    return await this.productsService.addProduct(user.sub, product);
   }
 
   @Mutation(() => ProductType)
@@ -85,7 +85,17 @@ export class ProductsResolver {
       throw new UnauthorizedException('You do not have the permission to archive a product');
     }
 
-    return this.productsService.archiveProduct(user.sub, productId);
+    const product = await this.productsService.productById(productId);
+
+    if (!product) {
+      throw new NotFoundException(`product with id ${productId} does not exists`);
+    }
+
+    if (user.sub !== product.userId) {
+      throw new UnauthorizedException('product is not yours to archive');
+    }
+
+    return await this.productsService.archiveProduct(productId);
   }
 
   @Mutation(() => ProductType)
@@ -97,7 +107,17 @@ export class ProductsResolver {
       throw new UnauthorizedException('You do not have the permission to restore a product');
     }
 
-    return this.productsService.restoreProduct(user.sub, productId);
+    const product = await this.productsService.productById(productId);
+
+    if (!product) {
+      throw new NotFoundException(`product with id ${productId} does not exists`);
+    }
+
+    if (user.sub !== product.userId) {
+      throw new UnauthorizedException('product is not yours to restore');
+    }
+
+    return this.productsService.restoreProduct(productId);
   }
 
   @Mutation(() => ProductType)
@@ -109,33 +129,40 @@ export class ProductsResolver {
       throw new UnauthorizedException('You do not have the permission to delete a product');
     }
 
+    const product = await this.productsService.productById(productId);
+
+    if (!product) {
+      throw new NotFoundException(`product with id ${productId} does not exists`);
+    }
+
     return this.productsService.deleteProduct(productId);
   }
 
   @Mutation(() => ProductType)
-  async updateStatus(
+  async updateProduct(
     @Context('user') user: IUser,
     @Args('productId') productId: string,
-    @Args('status') status: string
+    @Args('update') update: ProductInput
   ): Promise<ProductType> {
     if (!user || !user.permissions.includes('approve_changes:products')) {
-      throw new UnauthorizedException('You do not have the permission to update product status');
+      throw new UnauthorizedException('You do not have the permission to update product');
     }
 
-    pubSub.publish('productUpdated', {
-      productUpdated: `Very Much Updated to ${status}`,
-      productId
-    });
+    const product = await this.productsService.productById(productId);
 
-    return this.productsService.updateStatus(productId, status);
+    if (!product) {
+      throw new NotFoundException(`product with id ${productId} does not exists`);
+    }
+
+    return this.productsService.updateProduct(productId, update);
   }
 
-  @Subscription(() => String, {
-    filter: (payload, variables) => payload.productId === variables.productId
-  })
-  async productUpdated(
-    @Args('productId') productId: string
-  ) {
-    return pubSub.asyncIterator('productUpdated');
-  }
+  // @Subscription(() => ProductType, {
+  //   filter: (payload, variables) => payload.productId === variables.productId
+  // })
+  // async productUpdated(
+  //   @Args('productId') productId: string
+  // ) {
+  //   return pubSub.asyncIterator('productUpdated');
+  // }
 }
